adding support for open api autogen
